Minor fixes

+++ branches/prototype/xll/xlfoperimpl4.cpp
@@ -7,7 +7,7 @@ long XlfOperImpl4::strlen(const XlfOper &xlfOper) const {
 
     XLOPER xStr;
     Excel(xlCoerce, &xStr, 2, xlfOper.lpxloper4_, TempInt(xltypeStr));
-    long ret = xStr.val.str[0];
+    long ret = (BYTE)xStr.val.str[0];
     Excel(xlFree, 0, 1, &xStr);
     return ret;
